assets: css: site: standardization of images
This commit is a solution for incorrect settings of all sizes
in the post images, as well as handling the responsiveness of the images as well.

step-by-step problem: when opening any post with a large image,
the image size breaks the structure of the site horizontally, invading
the space of other elements and a scroll bar is created to navigate through the image.

Signed-off-by: Aquila Macedo <aquila0dayprofissional@gmail.com>